Extract prepare module from vm.
